service/requirements.txt
fastapi
uvicorn
llama-index
llama-index-llms-anthropic
#llama-index-embeddings-huggingface
google-cloud-storage